8.2 fix execution order
