+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/rest/UserApiTest.java
@@ -2,7 +2,6 @@ package guru.qa.niffler.test.rest;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.empty;
-import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.hasSize;
 import guru.qa.niffler.data.entity.userdata.UserEntity;
 import guru.qa.niffler.jupiter.annotation.meta.User;
@@ -17,17 +16,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
-import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.parallel.Execution;
-import org.junit.jupiter.api.parallel.ExecutionMode;
-import org.junit.jupiter.api.parallel.Isolated;
 
-@Isolated
-@Execution(ExecutionMode.SAME_THREAD)
-@TestMethodOrder(OrderAnnotation.class)
+
 @Epic("API")
 @Feature("User management")
 @Story("Friends")
@@ -80,7 +71,6 @@ public class UserApiTest {
     userClient.createFriends(user1, 1);
   }
 
-  @Order(1)
   @User
   @Test
   @DisplayName("API: Should returns empty list when search query doesn't match any user")
@@ -89,7 +79,6 @@ public class UserApiTest {
     assertThat(allUsers, empty());
   }
 
-  @Order(2)
   @User
   @Test
   @DisplayName("API: finds user when searching by 'mouse'")
@@ -99,12 +88,4 @@ public class UserApiTest {
     assertThat(allUsers, hasSize(1));
   }
 
-  @Order(Integer.MAX_VALUE)
-  @User
-  @Test
-  @DisplayName("API: Should return all users")
-  public void shouldReturnAllExistingUsersFromApi(UserJson user) {
-    List<UserJson> allUsers = userApiClient.getAllUsers(user.username());
-    assertThat(allUsers.size(), greaterThan(1));
-  }
 }
