+++ niffler-e-2-e-tests/src/test/resources/junit-platform.properties
@@ -1,6 +1,6 @@
 junit.jupiter.execution.parallel.enabled=true
 junit.jupiter.execution.parallel.mode.default=concurrent
-junit.jupiter.execution.parallel.mode.classes.default=concurrent
+junit.jupiter.execution.parallel.mode.classes.default=same_thread
 junit.jupiter.execution.parallel.config.strategy=fixed
 junit.jupiter.execution.parallel.config.fixed.parallelism=4
 junit.jupiter.execution.parallel.config.fixed.max-pool-size=4
